Feat: 트래커 뷰 컴포넌트 분리
