+++ lib/page/my_profile/my_profile_page.dart
@@ -6,6 +6,7 @@ import 'package:provider/provider.dart';
 
 import '../../model/class/badge.dart' as BadgeClass;
 import '../../model/class/user.dart';
+import '../../widget/tracker_view.dart';
 import '../common/Status.dart';
 import './my_profile_page_model.dart';
 import 'package:flutter/material.dart';
@@ -43,6 +44,9 @@ class MyProfileView extends StatelessWidget {
       return Loading(context: context);
     }
 
+    List<int> tracker = context.read<MyProfileViewModel>().tracker;
+    int postCount = context.read<MyProfileViewModel>().user.postCount;
+
     return ListView(
       physics: const ScrollPhysics(),
       children: [
@@ -51,7 +55,10 @@ class MyProfileView extends StatelessWidget {
         SizedBox(
           height: 16,
         ),
-        TrackerView(),
+        TrackerView(
+          tracker: tracker,
+          postCount: postCount,
+        ),
         SizedBox(
           height: 16,
         ),
@@ -108,47 +115,6 @@ class BadgeView extends StatelessWidget {
           ),
         ),
       ),
-    );
-  }
-}
-
-class TrackerView extends StatelessWidget {
-  const TrackerView({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    List<int> tracker = context.read<MyProfileViewModel>().tracker;
-
-    return Column(
-      children: [
-        Align(
-          alignment: Alignment.centerLeft,
-          child: Text(
-            'total: ${tracker.length} 퀘스트 완료',
-            style: const TextStyle(
-              fontSize: 14,
-            ),
-          ),
-        ),
-        const SizedBox(
-          height: 8,
-        ),
-        GridView.count(
-          crossAxisCount: 12,
-          crossAxisSpacing: 4,
-          mainAxisSpacing: 4,
-          shrinkWrap: true,
-          physics: const NeverScrollableScrollPhysics(),
-          children: List.generate(60, (index) {
-            return Container(
-              color: Color.fromRGBO(
-                  35, 236, 116, (tracker[index].toDouble() * 0.06) + 0.1),
-            );
-          }),
-        ),
-      ],
     );
   }
 }
